handling error finished

+++ src/intro-page/intro-page.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import {
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateIntroPageDto } from './dto/create-intro-page.dto';
 import { UpdateIntroPageDto } from './dto/update-intro-page.dto';
 import { IntroPage, IntroPageDocument } from './schemas/intro_page_schema';
@@ -31,28 +35,43 @@ export class IntroPageService {
 
   async findAll(shopId?: Types.ObjectId) {
     try {
+      const shop = await this.shopModel.findById(shopId);
+      if (!shop) {
+        throw new NotFoundException('this shop not found');
+      }
       const introPage = await this.introPageModel.find({ shopId });
       return introPage;
     } catch (error) {
-      throw new InternalServerErrorException(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      } else {
+        throw new InternalServerErrorException(error);
+      }
     }
   }
 
   async findOne(id: string) {
     try {
-      const introPage = await this.introPageModel.findById(id).catch((err) => {
-        console.log(err);
-        throw new InternalServerErrorException(err);
-      });
+      const introPage = await this.introPageModel.findById(id);
+      if (!introPage) {
+        throw new NotFoundException('this page not found');
+      }
       return introPage;
     } catch (error) {
-      console.log(error);
-      throw new InternalServerErrorException(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      } else {
+        throw new InternalServerErrorException(error);
+      }
     }
   }
 
   async update(id: string, updateIntroPageDto: UpdateIntroPageDto) {
     try {
+      const introPageExist = await this.introPageModel.findById(id);
+      if (!introPageExist) {
+        throw new NotFoundException('this page not found');
+      }
       const introPage = await this.introPageModel.findByIdAndUpdate(
         id,
         updateIntroPageDto,
@@ -60,12 +79,20 @@ export class IntroPageService {
       );
       return introPage;
     } catch (error) {
-      throw new InternalServerErrorException(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      } else {
+        throw new InternalServerErrorException(error);
+      }
     }
   }
 
   async remove(id: string) {
     try {
+      const introPageExist = await this.introPageModel.findById(id);
+      if (!introPageExist) {
+        throw new NotFoundException('this page not found');
+      }
       const introPage = await this.introPageModel.findByIdAndDelete(id);
       const shop = await this.shopModel.findById(introPage.shopId);
       for (let i = 0; i < shop.introPages.length; i++) {
@@ -74,8 +101,11 @@ export class IntroPageService {
       await shop.save();
       return 'Intro page has been deleted successfully';
     } catch (error) {
-      console.log(error);
-      throw new InternalServerErrorException(error);
+      if (error instanceof NotFoundException) {
+        throw error;
+      } else {
+        throw new InternalServerErrorException(error);
+      }
     }
   }
 }
